* Split some image utils into `panoptes.utils.images.misc`
