+++ tests/images/test_image_utils.py
@@ -5,28 +5,28 @@ import numpy as np
 import pytest
 from astropy.nddata import Cutout2D
 
+import panoptes.utils.images.misc
 from panoptes.utils import error
-from panoptes.utils import images as img_utils
 
 
 def test_crop_data():
     ones = np.ones((201, 201))
     assert ones.sum() == 40401.
 
-    cropped01 = img_utils.crop_data(ones)  # False to exercise coverage.
+    cropped01 = panoptes.utils.images.misc.crop_data(ones)  # False to exercise coverage.
     assert cropped01.sum() == 40000.
 
-    cropped02 = img_utils.crop_data(ones, box_width=10)
+    cropped02 = panoptes.utils.images.misc.crop_data(ones, box_width=10)
     assert cropped02.sum() == 100.
 
-    cropped03 = img_utils.crop_data(ones, box_width=6, center=(50, 50))
+    cropped03 = panoptes.utils.images.misc.crop_data(ones, box_width=6, center=(50, 50))
     assert cropped03.sum() == 36.
 
     # Test the Cutout2D object
-    cropped04 = img_utils.crop_data(ones,
-                                    box_width=20,
-                                    center=(50, 50),
-                                    data_only=False)
+    cropped04 = panoptes.utils.images.misc.crop_data(ones,
+                                                     box_width=20,
+                                                     center=(50, 50),
+                                                     data_only=False)
     assert isinstance(cropped04, Cutout2D)
     assert cropped04.position_original == (50, 50)
 
@@ -41,7 +41,7 @@ def test_make_pretty_image(solved_fits_file, tiny_fits_file, save_environ):
 
         # Can't operate on a non-existent files.
         with pytest.warns(UserWarning, match="File doesn't exist"):
-            assert not img_utils.make_pretty_image('Foobar')
+            assert not panoptes.utils.images.misc.make_pretty_image('Foobar')
 
         # Can handle the fits file, and creating the images dir for linking
         # the latest image.
@@ -50,7 +50,7 @@ def test_make_pretty_image(solved_fits_file, tiny_fits_file, save_environ):
         os.makedirs(imgdir, exist_ok=True)
 
         link_path = os.path.join(tmpdir, 'latest.jpg')
-        pretty = img_utils.make_pretty_image(solved_fits_file, link_path=link_path)
+        pretty = panoptes.utils.images.misc.make_pretty_image(solved_fits_file, link_path=link_path)
         assert pretty.exists()
         assert pretty.is_file()
         assert os.path.isdir(imgdir)
@@ -59,7 +59,7 @@ def test_make_pretty_image(solved_fits_file, tiny_fits_file, save_environ):
         os.rmdir(imgdir)
 
         # Try again, but without link_path.
-        pretty = img_utils.make_pretty_image(tiny_fits_file, title='some text')
+        pretty = panoptes.utils.images.misc.make_pretty_image(tiny_fits_file, title='some text')
         assert pretty.exists()
         assert pretty.is_file()
         assert not os.path.isdir(imgdir)
@@ -71,21 +71,21 @@ def test_make_pretty_image_cr2_fail():
         with open(tmpfile, 'w') as f:
             f.write('not an image file')
         with pytest.raises(error.InvalidSystemCommand):
-            img_utils.make_pretty_image(tmpfile, title='some text')
+            panoptes.utils.images.misc.make_pretty_image(tmpfile, title='some text')
         with pytest.raises(error.AlreadyExists):
-            img_utils.make_pretty_image(tmpfile)
+            panoptes.utils.images.misc.make_pretty_image(tmpfile)
 
-        no_image = img_utils.make_pretty_image('not-a-file')
+        no_image = panoptes.utils.images.misc.make_pretty_image('not-a-file')
         assert no_image is None
 
 
 def test_make_pretty_image_cr2(cr2_file, tmpdir):
     link_path = str(tmpdir.mkdir('images').join('latest.jpg'))
     print(f'link_path: {link_path} cr2_file: {cr2_file}')
-    pretty_path = img_utils.make_pretty_image(cr2_file,
-                                              title='CR2 Test',
-                                              link_path=link_path,
-                                              remove_cr2=True)
+    pretty_path = panoptes.utils.images.misc.make_pretty_image(cr2_file,
+                                                               title='CR2 Test',
+                                                               link_path=link_path,
+                                                               remove_cr2=True)
 
     assert pretty_path.exists()
     assert pretty_path.as_posix() == link_path
